fix(pgvector): fix Collection not found error
When doing python init_database.py --recreate-vs with pgvector as
vector store, it would occur "Collection not found error". And all
existing kb docs would not be added to pgvector.

This is happened because when recreating vs, kb.clear_vs() would be
called to do cleaning. At this time, collection is deleted. It will
not be created automatically.

Fix this by re-creating the collection manually just after deleting
it.

Fixes: #1288

Signed-off-by: Kyle Zhang <zhangk1985@gmail.com>